created backup & restore as well as initialization scripts. updated documentation
The scripts will setup collections for the project with schema validation in place, as well as a DMLUser and a superadmin

+++ SkiServiceAPI/Common/Generics/GenericController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Bson;
 using SkiServiceAPI.Interfaces;
+using SkiServiceModels.BSON.Interfaces.Base;
 using SkiServiceModels.Enums;
 using SkiServiceModels.Interfaces;
 
@@ -16,17 +17,16 @@ namespace SkiServiceAPI.Common
     /// <typeparam name="TCreate">Create DTO</typeparam>
     [ApiController]
     [Route("api/[controller]")]
-    public class GenericController<T, TResponseBase, TResponseAdmin, TUpdate, TCreate> : ControllerBase
-        where T : class, IGenericBSONModel
-        where TResponseBase : class
-        where TResponseAdmin : class, TResponseBase
+    public class GenericController<T, TResponse, TUpdate, TCreate> : ControllerBase
+        where T : class, IModel
+        where TResponse : class
         where TUpdate : class
         where TCreate : class
     {
 
-        private readonly IBaseService<T, TResponseBase, TResponseAdmin, TUpdate, TCreate> _service;
+        private readonly IBaseService<T, TResponse, TUpdate, TCreate> _service;
 
-        public GenericController(IBaseService<T, TResponseBase, TResponseAdmin, TUpdate, TCreate> service)
+        public GenericController(IBaseService<T, TResponse, TUpdate, TCreate> service)
         {
             _service = service;
         }
@@ -36,6 +36,7 @@ namespace SkiServiceAPI.Common
         /// </summary>
         /// <returns>TDestination mapped from T</returns>
         [HttpGet]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public virtual async Task<IActionResult> GetAll()
         {
@@ -49,6 +50,7 @@ namespace SkiServiceAPI.Common
         /// <param name="id">Id</param>
         /// <returns>TDestination mapped from T</returns>
         [HttpGet("{id}")]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public virtual async Task<IActionResult> Get(string id)
         {
@@ -71,17 +73,19 @@ namespace SkiServiceAPI.Common
             return result.IsSuccess ? Ok(result.Result) : BadRequest(result.Error);
         }
 
-        /// <summary>
+        /// <summary>   
         /// Update entity
         /// </summary>
         /// <param name="entity">Entity Data</param>
         /// <returns>TDestination mapped from T</returns>
         [HttpPut("{id}")]
-        [Authorize(Roles = nameof(RoleNames.SuperAdmin))]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public virtual async Task<IActionResult> Update(string id, [FromBody] TUpdate entity)
         {
+            if(!await _service.IsOwnerOrAdmin(ObjectId.Parse(id))) return Forbid();
             var result = await _service.UpdateAsync(ObjectId.Parse(id), entity);
             return result.IsSuccess ? Ok(result.Result) : NotFound(result.Error);
         }
@@ -92,11 +96,13 @@ namespace SkiServiceAPI.Common
         /// <param name="id">Id</param>
         /// <returns>TDestination mapped from T</returns>
         [HttpDelete("{id}")]
-        [Authorize(Roles = nameof(RoleNames.SuperAdmin))]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public virtual async Task<IActionResult> Delete(string id)
         {
+            if (!await _service.IsOwnerOrAdmin(ObjectId.Parse(id))) return Forbid();
             var result = await _service.DeleteAsync(ObjectId.Parse(id));
             return result.IsSuccess ? Ok(result.Result) : NotFound(result.Error);
         }
